Cast encoder inputs to the live weight dtype
The ECG encoders hardcoded ecg.float(), which breaks under FSDP's
mixed-precision policy: compute parameters are bfloat16, so the first
linear/conv after the cast saw mixed dtypes (caught by the 2-GPU
FSDP + llava distributed case). Cast to the embed/stem weight's
current dtype instead — identical behavior single-GPU (fp32 params),
correct under any mixed-precision policy. Same principle the
connectors already follow.

Verified: 2-GPU FSDP mlp_llava+st_mem trains; single-GPU model suite
11/11 unchanged.

+++ src/models/encoders/merl.py
@@ -36,5 +36,5 @@ class Merl(nn.Module):
         self.pool = nn.AdaptiveAvgPool1d(num_tokens)
 
     def forward(self, ecg):
-        x = self.stages(self.stem(ecg.float()))
+        x = self.stages(self.stem(ecg.to(self.stem[0].weight.dtype)))
         return self.pool(x).transpose(1, 2)  # (B, num_tokens, 2048)
